✨ Refactor: Timer Component
